refactor: rename tables to models for convention purposes
