+++ backend/api/v1/mcc/routes/users.py
@@ -1,7 +1,7 @@
 from api.v1.mcc.schemas.requests import UpdateUserRequest
 from api.v1.mcc.schemas.responses import UserInformationResponse
 from data.data_wrappers.wrappers import MCCUsersWrapper
-from data.tables.mcc_user_tables import MCCUsers
+from data.models.mcc_user_models import MCCUsers
 from fastapi import APIRouter, Depends
 from fastapi.exceptions import HTTPException
 from mcc_keycloak.client import keycloak
