+++ backend/data/data_wrappers/abstract_wrapper.py
@@ -5,7 +5,7 @@ from uuid import UUID
 from sqlmodel import select
 
 from data.database.engine import get_db_session
-from data.tables.base_model import BaseSQLModel
+from data.models.base_model import BaseSQLModel
 
 T = TypeVar("T", bound=BaseSQLModel)
 PK = TypeVar("PK", int, UUID)
