+++ backend/data/database/engine.py
@@ -4,9 +4,9 @@ from config.config import settings
 from sqlalchemy import Engine
 from sqlmodel import Session, create_engine, text
 
-from data.tables.aro_user_tables import ARO_USER_SCHEMA_NAME
-from data.tables.main_tables import MAIN_SCHEMA_NAME
-from data.tables.transactional_tables import TRANSACTIONAL_SCHEMA_NAME
+from data.models.aro_user_models import ARO_USER_SCHEMA_NAME
+from data.models.main_models import MAIN_SCHEMA_NAME
+from data.models.transactional_models import TRANSACTIONAL_SCHEMA_NAME
 
 
 @lru_cache(maxsize=1)
