+++ backend/data/models/main_models.py
@@ -5,7 +5,7 @@ from pydantic import model_validator
 from sqlalchemy import Integer
 from sqlmodel import Field
 
-from data.tables.base_model import BaseSQLModel
+from data.models.base_model import BaseSQLModel
 
 # Schema information
 MAIN_SCHEMA_NAME: Final[str] = "main"
