+++ backend/mcc_keycloak/client.py
@@ -5,7 +5,7 @@ from uuid import UUID
 from config.config import settings
 from config.keycloak_config import KeycloakConfig
 from data.data_wrappers.wrappers import MCCUsersWrapper
-from data.tables.mcc_user_tables import MCCUsers
+from data.models.mcc_user_models import MCCUsers
 from fastapi import Depends, Request, status
 from fastapi.exceptions import HTTPException
 from jwcrypto.jwt import JWTExpired
