+++ backend/python_test/test_tables/test_aro_user.py
@@ -1,5 +1,5 @@
 import pytest
-from data.tables.aro_user_tables import AROUsers
+from data.models.aro_user_models import AROUsers
 from pydantic import ValidationError
 from sqlmodel import Session, select
